+++ src/commands/_common/parseCommand.ts
@@ -1,5 +1,5 @@
 import { spaceTrim } from 'spacetrim';
-import { ParsingError } from '../../errors/ParsingError';
+import { ParseError } from '../../errors/ParseError';
 import type { string_markdown, string_markdown_text } from '../../types/typeAliases';
 import { removeMarkdownFormatting } from '../../utils/markdown/removeMarkdownFormatting';
 import { normalizeTo_SCREAMING_CASE } from '../../utils/normalization/normalizeTo_SCREAMING_CASE';
@@ -15,13 +15,13 @@ import type { CommandUsagePlace } from './types/CommandUsagePlaces';
  * Parses one line of ul/ol to command
  *
  * @returns parsed command object
- * @throws {ParsingError} if the command is invalid
+ * @throws {ParseError} if the command is invalid
  *
  * @private within the pipelineStringToJson
  */
 export function parseCommand(raw: string_markdown_text, usagePlace: CommandUsagePlace): Command {
     if (raw.includes('\n') || raw.includes('\r')) {
-        throw new ParsingError('Command can not contain new line characters' /* <- TODO: [🚞] */);
+        throw new ParseError('Command can not contain new line characters' /* <- TODO: [🚞] */);
     }
 
     let normalized = raw.trim();
@@ -63,7 +63,7 @@ export function parseCommand(raw: string_markdown_text, usagePlace: CommandUsage
         .map((item) => item.trim());
 
     if (items.length === 0 || items[0] === '') {
-        throw new ParsingError(
+        throw new ParseError(
             spaceTrim(
                 (block) =>
                     `
@@ -112,7 +112,7 @@ export function parseCommand(raw: string_markdown_text, usagePlace: CommandUsage
         }
     }
 
-    throw new ParsingError(
+    throw new ParseError(
         spaceTrim(
             (block) =>
                 `
@@ -167,16 +167,16 @@ function parseCommandVariant(input: CommandParserInput & { commandNameRaw: strin
             try {
                 return parse({ usagePlace, raw, rawArgs, normalized, args }) as Command; // <- Note: [🦦]
             } catch (error) {
-                if (!(error instanceof ParsingError)) {
+                if (!(error instanceof ParseError)) {
                     throw error;
                 }
 
-                throw new ParsingError(
+                throw new ParseError(
                     spaceTrim(
                         (block) =>
                             `
                               Invalid ${commandName} command:
-                              ${block((error as ParsingError).message)}
+                              ${block((error as ParseError).message)}
 
                               - ${raw}