+++ src/conversion/utils/extractVariables.ts
@@ -1,12 +1,12 @@
 import { spaceTrim } from 'spacetrim';
-import { ParsingError } from '../../errors/ParsingError';
+import { ParseError } from '../../errors/ParseError';
 import type { string_javascript, string_javascript_name } from '../../types/typeAliases';
 /**
  * Parses the given script and returns the list of all used variables that are not defined in the script
  *
  * @param script from which to extract the variables
  * @returns the list of variable names
- * @throws {ParsingError} if the script is invalid
+ * @throws {ParseError} if the script is invalid
  * @public exported from `@promptbook/utils`
  */
 export function extractVariables(script: string_javascript): Set<string_javascript_name> {
@@ -44,7 +44,7 @@ export function extractVariables(script: string_javascript): Set<string_javascri
             throw error;
         }
 
-        throw new ParsingError(
+        throw new ParseError(
             spaceTrim(
                 (block) => `
                     Can not extract variables from the script